유저 활동 관련 서비스 분리

+++ module-activity/src/main/java/Project/SocialCommerce/service/PostService.java
@@ -7,6 +7,7 @@ import Project.SocialCommerce.dto.PostResponseDto;
 import Project.SocialCommerce.dto.UserResponseDto;
 import Project.SocialCommerce.model.Comment;
 import Project.SocialCommerce.model.Post;
+import Project.SocialCommerce.repository.CommentRepository;
 import Project.SocialCommerce.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -19,7 +20,7 @@ public class PostService {
 
     private final PostRepository postRepository;
     private final UserClient userClient;
-//    private final UserRepository userRepository;
+    private final CommentRepository commentRepository;
 //    private final ActivityRepository activityRepository;
 
 
@@ -54,32 +55,34 @@ public class PostService {
         return new PostResponseDto(post.get());
     }
 
-    public void editPost(EditPostRequestDto editPostRequestDto, String email) {
+    public void editPost(EditPostRequestDto editPostRequestDto, String jwt) {
         Optional<Post> editPostOpt = postRepository.findById(editPostRequestDto.getPostId());
         if (editPostOpt.isEmpty()){
             throw new IllegalArgumentException("없는 게시물 입니다.");
         }
         Post editPost = editPostOpt.get();
 
-//        if (!editPost.getUser().getEmail().equals(email)) {
-//            throw new IllegalArgumentException("권한이 없는 사용자 입니다.");
-//        }
+        if (!editPost.getEmail().equals(userClient.findByJwt(jwt).getEmail())) {
+            throw new IllegalArgumentException("권한이 없는 사용자 입니다.");
+        }
 
         editPost.setContent(editPostRequestDto.getContent());
 
         postRepository.save(editPost);
     }
 
-    public void delPost(EditPostRequestDto delPostRequestDto, String email) {
+    public void delPost(EditPostRequestDto delPostRequestDto, String jwt) {
         Optional<Post> delPostOPt = postRepository.findById(delPostRequestDto.getPostId());
         if (delPostOPt.isEmpty()) {
             throw new IllegalArgumentException("없는 게시물 입니다.");
         }
         Post delPost = delPostOPt.get();
 
-//        if (!delPost.getUser().getEmail().equals(email)) {
-//            throw new IllegalArgumentException("권한이 없는 사용자 입니다.");
-//        }
+        if (!delPost.getEmail().equals(userClient.findByJwt(jwt).getEmail())) {
+            throw new IllegalArgumentException("권한이 없는 사용자 입니다.");
+        }
+
+        commentRepository.deleteAll(delPost.getComments());
 
         postRepository.delete(delPost);
     }
